refactor: 분석 채팅 훅 경계 정리
